Diff-in-diff in four different ways
Estimation by a standard regression with state and year fixed effects, a standard regression but weighted with each state in each year having the same weight, regression with the mean of all variables, and regression with the mean of all variables with each state in each year weighted to be representative of the sample. Event study looking at the effect of the merit aid program.